Moved ASR logic to view model
